fix: impl std Error for StartError
Also extract StartError code into a standalone file.
