+++ Cargo.toml
@@ -17,6 +17,7 @@ build = "build.rs"
 log = "0.4.17"
 
 [dev-dependencies]
+anyhow = "1.0.71"
 env_logger = "0.10.0"
 pretty_assertions = "1.3.0"
 ureq = "2.6.2"
